feat(CycleTimer): Refactor CycleTimer component with Default and Small variants
