+++ src/components/CycleTimer/CycleTimer.ts
@@ -1,0 +1,6 @@
+export { CycleTimer } from "./index"
+export type {
+  CycleTimerControls,
+  CycleTimerProps,
+  CycleTimerState,
+} from "./types"
